CCMSPUI-454: Added integration tests
