+++ src/integrationTest/java/uk/gov/laa/ccms/caab/client/SoaApiClientIntegrationTest.java
@@ -21,7 +21,6 @@ import org.springframework.test.context.DynamicPropertySource;
 import reactor.core.publisher.Mono;
 import uk.gov.laa.ccms.caab.AbstractIntegrationTest;
 import uk.gov.laa.ccms.caab.bean.CaseSearchCriteria;
-import uk.gov.laa.ccms.caab.bean.NotificationSearchCriteria;
 import uk.gov.laa.ccms.soa.gateway.model.BaseClient;
 import uk.gov.laa.ccms.soa.gateway.model.CaseDetails;
 import uk.gov.laa.ccms.soa.gateway.model.CaseSummary;
@@ -139,34 +138,6 @@ public class SoaApiClientIntegrationTest extends AbstractIntegrationTest {
     ClientDetail response = clientMono.block();
 
     assertEquals(clientDetail, response);
-  }
-
-  @Test
-  public void testGetNotifications_returnsData() throws JsonProcessingException {
-    Notifications notifications = buildNotifications();
-    String notificationsJson = objectMapper.writeValueAsString(notifications);
-
-    NotificationSearchCriteria criteria = new NotificationSearchCriteria();
-    criteria.setAssignedToUserId("testUserId");
-
-    criteria.setLoginId("testUserId");
-    criteria.setUserType("testUserType");
-    int page = 10;
-    int size = 10;
-
-    wiremock.stubFor(get(String.format("/notifications?assigned-to-user-id=%s&include-closed=%s&page=%s&" +
-            "size=%s",
-        criteria.getAssignedToUserId(),
-        criteria.isIncludeClosed(),
-        page,
-        size))
-        .withHeader(SOA_GATEWAY_USER_LOGIN_ID, equalTo(criteria.getLoginId()))
-        .withHeader(SOA_GATEWAY_USER_ROLE, equalTo(criteria.getUserType()))
-        .willReturn(okJson(notificationsJson)));
-    Mono<Notifications> notificationsMono =
-        soaApiClient.getNotifications(criteria, page, size);
-    Notifications response = notificationsMono.block();
-    assertEquals(notifications, response);
   }
 
   @Test
